good

+++ bridge/src/main/java/org/simulspeak/bridge/configuration/BridgeConfig.java
@@ -10,12 +10,12 @@ public class BridgeConfig {
 
     public static final Long ERROR_VIDEO_ID = -1L;
 
-    public static final Long APPLY_VIDEO_ENUM = 0L;
+    public static final int APPLY_VIDEO_ENUM = 0;
 
-    public static final Long APPLY_IMAGE_ENUM = 1L;
+    public static final int APPLY_IMAGE_ENUM = 1;
 
-    public static final Long APPLY_COMMENT_ENUM = 2L;
+    public static final int APPLY_COMMENT_ENUM = 2;
 
-    public static final Long APPLY_AUDIO_ENUM = 3L;
+    public static final int APPLY_AUDIO_ENUM = 3;
     
 }
